+++ src/styles/three-canvas-style.tsx
@@ -3,10 +3,12 @@ import { css } from '@emotion/core';
 export const threeCanvasStyle = css`
   flex: 1;
   width: 100vw;
+
   div {
     position: absolute !important;
     top: 0;
     left: 0;
+    z-index: -1;
   }
 
   @media screen and (max-width: 1100px) {
